requirements.txt
Django==4.0
petl==1.7.4
requests==2.26.0
marshmallow==3.14.1
python-dateutil==2.8.2
parameterized==0.8.1
